+++ rule_engine/rcm/src/main/java/com/tricon/rcm/dto/RcmPatientStatementDto.java
@@ -16,6 +16,6 @@ public class RcmPatientStatementDto {
 	private String remarks;
 	private String statementNotes;
 	private String balanceSheetLink;
-	private byte buttonType;
+	private int buttonType;
 	private int sectionId;
 }
